Refactor `alloc/sync.rs` into separate files
- `alloc::sync::Arc` now lives in `libraries/alloc/sync/arc.rs`
- `alloc::sync::Weak` now lives in `libraries/alloc/sync/weak.rs`
- tests now live in `alloc/tests/arc.rs`
