+++ library/alloc/src/sync/arc.rs
@@ -1,9 +1,3 @@
-#![stable(feature = "rust1", since = "1.0.0")]
-
-//! Thread-safe reference-counting pointers.
-//!
-//! See the [`Arc<T>`][Arc] documentation for more details.
-
 use core::any::Any;
 use core::borrow;
 use core::cmp::Ordering;
@@ -29,10 +23,8 @@ use crate::borrow::{Cow, ToOwned};
 use crate::boxed::Box;
 use crate::rc::is_dangling;
 use crate::string::String;
+use crate::sync::Weak;
 use crate::vec::Vec;
-
-#[cfg(test)]
-mod tests;
 
 /// A soft limit on the amount of references that may be made to an `Arc`.
 ///
@@ -239,7 +231,7 @@ impl<T: ?Sized + Unsize<U>, U: ?Sized> CoerceUnsized<Arc<U>> for Arc<T> {}
 impl<T: ?Sized + Unsize<U>, U: ?Sized> DispatchFromDyn<Arc<U>> for Arc<T> {}
 
 impl<T: ?Sized> Arc<T> {
-    fn from_inner(ptr: NonNull<ArcInner<T>>) -> Self {
+    pub(super) fn from_inner(ptr: NonNull<ArcInner<T>>) -> Self {
         Self { ptr, phantom: PhantomData }
     }
 
@@ -248,67 +240,19 @@ impl<T: ?Sized> Arc<T> {
     }
 }
 
-/// `Weak` is a version of [`Arc`] that holds a non-owning reference to the
-/// managed allocation. The allocation is accessed by calling [`upgrade`] on the `Weak`
-/// pointer, which returns an [`Option`]`<`[`Arc`]`<T>>`.
-///
-/// Since a `Weak` reference does not count towards ownership, it will not
-/// prevent the value stored in the allocation from being dropped, and `Weak` itself makes no
-/// guarantees about the value still being present. Thus it may return [`None`]
-/// when [`upgrade`]d. Note however that a `Weak` reference *does* prevent the allocation
-/// itself (the backing store) from being deallocated.
-///
-/// A `Weak` pointer is useful for keeping a temporary reference to the allocation
-/// managed by [`Arc`] without preventing its inner value from being dropped. It is also used to
-/// prevent circular references between [`Arc`] pointers, since mutual owning references
-/// would never allow either [`Arc`] to be dropped. For example, a tree could
-/// have strong [`Arc`] pointers from parent nodes to children, and `Weak`
-/// pointers from children back to their parents.
-///
-/// The typical way to obtain a `Weak` pointer is to call [`Arc::downgrade`].
-///
-/// [`upgrade`]: Weak::upgrade
-#[stable(feature = "arc_weak", since = "1.4.0")]
-pub struct Weak<T: ?Sized> {
-    // This is a `NonNull` to allow optimizing the size of this type in enums,
-    // but it is not necessarily a valid pointer.
-    // `Weak::new` sets this to `usize::MAX` so that it doesn’t need
-    // to allocate space on the heap.  That's not a value a real pointer
-    // will ever have because RcBox has alignment at least 2.
-    // This is only possible when `T: Sized`; unsized `T` never dangle.
-    ptr: NonNull<ArcInner<T>>,
-}
-
-#[stable(feature = "arc_weak", since = "1.4.0")]
-unsafe impl<T: ?Sized + Sync + Send> Send for Weak<T> {}
-#[stable(feature = "arc_weak", since = "1.4.0")]
-unsafe impl<T: ?Sized + Sync + Send> Sync for Weak<T> {}
-
-#[unstable(feature = "coerce_unsized", issue = "27732")]
-impl<T: ?Sized + Unsize<U>, U: ?Sized> CoerceUnsized<Weak<U>> for Weak<T> {}
-#[unstable(feature = "dispatch_from_dyn", issue = "none")]
-impl<T: ?Sized + Unsize<U>, U: ?Sized> DispatchFromDyn<Weak<U>> for Weak<T> {}
-
-#[stable(feature = "arc_weak", since = "1.4.0")]
-impl<T: ?Sized + fmt::Debug> fmt::Debug for Weak<T> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "(Weak)")
-    }
-}
-
 // This is repr(C) to future-proof against possible field-reordering, which
 // would interfere with otherwise safe [into|from]_raw() of transmutable
 // inner types.
 #[repr(C)]
-struct ArcInner<T: ?Sized> {
-    strong: atomic::AtomicUsize,
+pub(super) struct ArcInner<T: ?Sized> {
+    pub(super) strong: atomic::AtomicUsize,
 
     // the value usize::MAX acts as a sentinel for temporarily "locking" the
     // ability to upgrade weak pointers or downgrade strong ones; this is used
     // to avoid races in `make_mut` and `get_mut`.
-    weak: atomic::AtomicUsize,
+    pub(super) weak: atomic::AtomicUsize,
 
-    data: T,
+    pub(super) data: T,
 }
 
 unsafe impl<T: ?Sized + Sync + Send> Send for ArcInner<T> {}
@@ -1322,7 +1266,7 @@ impl<T: Clone> Arc<T> {
     ///
     /// [clone]: Clone::clone
     /// [get_mut]: Arc::get_mut
-    /// [`Rc::make_mut`]: super::rc::Rc::make_mut
+    /// [`Rc::make_mut`]: crate::rc::Rc::make_mut
     ///
     /// # Examples
     ///
@@ -1604,434 +1548,6 @@ impl Arc<dyn Any + Send + Sync> {
             Ok(Arc::from_inner(ptr))
         } else {
             Err(self)
-        }
-    }
-}
-
-impl<T> Weak<T> {
-    /// Constructs a new `Weak<T>`, without allocating any memory.
-    /// Calling [`upgrade`] on the return value always gives [`None`].
-    ///
-    /// [`upgrade`]: Weak::upgrade
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use std::sync::Weak;
-    ///
-    /// let empty: Weak<i64> = Weak::new();
-    /// assert!(empty.upgrade().is_none());
-    /// ```
-    #[stable(feature = "downgraded_weak", since = "1.10.0")]
-    pub fn new() -> Weak<T> {
-        Weak { ptr: NonNull::new(usize::MAX as *mut ArcInner<T>).expect("MAX is not 0") }
-    }
-}
-
-/// Helper type to allow accessing the reference counts without
-/// making any assertions about the data field.
-struct WeakInner<'a> {
-    weak: &'a atomic::AtomicUsize,
-    strong: &'a atomic::AtomicUsize,
-}
-
-impl<T: ?Sized> Weak<T> {
-    /// Returns a raw pointer to the object `T` pointed to by this `Weak<T>`.
-    ///
-    /// The pointer is valid only if there are some strong references. The pointer may be dangling,
-    /// unaligned or even [`null`] otherwise.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use std::sync::Arc;
-    /// use std::ptr;
-    ///
-    /// let strong = Arc::new("hello".to_owned());
-    /// let weak = Arc::downgrade(&strong);
-    /// // Both point to the same object
-    /// assert!(ptr::eq(&*strong, weak.as_ptr()));
-    /// // The strong here keeps it alive, so we can still access the object.
-    /// assert_eq!("hello", unsafe { &*weak.as_ptr() });
-    ///
-    /// drop(strong);
-    /// // But not any more. We can do weak.as_ptr(), but accessing the pointer would lead to
-    /// // undefined behaviour.
-    /// // assert_eq!("hello", unsafe { &*weak.as_ptr() });
-    /// ```
-    ///
-    /// [`null`]: core::ptr::null
-    #[stable(feature = "weak_into_raw", since = "1.45.0")]
-    pub fn as_ptr(&self) -> *const T {
-        let ptr: *mut ArcInner<T> = NonNull::as_ptr(self.ptr);
-
-        if is_dangling(ptr) {
-            // If the pointer is dangling, we return the sentinel directly. This cannot be
-            // a valid payload address, as the payload is at least as aligned as ArcInner (usize).
-            ptr as *const T
-        } else {
-            // SAFETY: if is_dangling returns false, then the pointer is dereferencable.
-            // The payload may be dropped at this point, and we have to maintain provenance,
-            // so use raw pointer manipulation.
-            unsafe { ptr::addr_of_mut!((*ptr).data) }
-        }
-    }
-
-    /// Consumes the `Weak<T>` and turns it into a raw pointer.
-    ///
-    /// This converts the weak pointer into a raw pointer, while still preserving the ownership of
-    /// one weak reference (the weak count is not modified by this operation). It can be turned
-    /// back into the `Weak<T>` with [`from_raw`].
-    ///
-    /// The same restrictions of accessing the target of the pointer as with
-    /// [`as_ptr`] apply.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use std::sync::{Arc, Weak};
-    ///
-    /// let strong = Arc::new("hello".to_owned());
-    /// let weak = Arc::downgrade(&strong);
-    /// let raw = weak.into_raw();
-    ///
-    /// assert_eq!(1, Arc::weak_count(&strong));
-    /// assert_eq!("hello", unsafe { &*raw });
-    ///
-    /// drop(unsafe { Weak::from_raw(raw) });
-    /// assert_eq!(0, Arc::weak_count(&strong));
-    /// ```
-    ///
-    /// [`from_raw`]: Weak::from_raw
-    /// [`as_ptr`]: Weak::as_ptr
-    #[stable(feature = "weak_into_raw", since = "1.45.0")]
-    pub fn into_raw(self) -> *const T {
-        let result = self.as_ptr();
-        mem::forget(self);
-        result
-    }
-
-    /// Converts a raw pointer previously created by [`into_raw`] back into `Weak<T>`.
-    ///
-    /// This can be used to safely get a strong reference (by calling [`upgrade`]
-    /// later) or to deallocate the weak count by dropping the `Weak<T>`.
-    ///
-    /// It takes ownership of one weak reference (with the exception of pointers created by [`new`],
-    /// as these don't own anything; the method still works on them).
-    ///
-    /// # Safety
-    ///
-    /// The pointer must have originated from the [`into_raw`] and must still own its potential
-    /// weak reference.
-    ///
-    /// It is allowed for the strong count to be 0 at the time of calling this. Nevertheless, this
-    /// takes ownership of one weak reference currently represented as a raw pointer (the weak
-    /// count is not modified by this operation) and therefore it must be paired with a previous
-    /// call to [`into_raw`].
-    /// # Examples
-    ///
-    /// ```
-    /// use std::sync::{Arc, Weak};
-    ///
-    /// let strong = Arc::new("hello".to_owned());
-    ///
-    /// let raw_1 = Arc::downgrade(&strong).into_raw();
-    /// let raw_2 = Arc::downgrade(&strong).into_raw();
-    ///
-    /// assert_eq!(2, Arc::weak_count(&strong));
-    ///
-    /// assert_eq!("hello", &*unsafe { Weak::from_raw(raw_1) }.upgrade().unwrap());
-    /// assert_eq!(1, Arc::weak_count(&strong));
-    ///
-    /// drop(strong);
-    ///
-    /// // Decrement the last weak count.
-    /// assert!(unsafe { Weak::from_raw(raw_2) }.upgrade().is_none());
-    /// ```
-    ///
-    /// [`new`]: Weak::new
-    /// [`into_raw`]: Weak::into_raw
-    /// [`upgrade`]: Weak::upgrade
-    /// [`forget`]: std::mem::forget
-    #[stable(feature = "weak_into_raw", since = "1.45.0")]
-    pub unsafe fn from_raw(ptr: *const T) -> Self {
-        // See Weak::as_ptr for context on how the input pointer is derived.
-
-        let ptr = if is_dangling(ptr as *mut T) {
-            // This is a dangling Weak.
-            ptr as *mut ArcInner<T>
-        } else {
-            // Otherwise, we're guaranteed the pointer came from a nondangling Weak.
-            // SAFETY: data_offset is safe to call, as ptr references a real (potentially dropped) T.
-            let offset = unsafe { data_offset(ptr) };
-            // Thus, we reverse the offset to get the whole RcBox.
-            // SAFETY: the pointer originated from a Weak, so this offset is safe.
-            unsafe { (ptr as *mut ArcInner<T>).set_ptr_value((ptr as *mut u8).offset(-offset)) }
-        };
-
-        // SAFETY: we now have recovered the original Weak pointer, so can create the Weak.
-        Weak { ptr: unsafe { NonNull::new_unchecked(ptr) } }
-    }
-}
-
-impl<T: ?Sized> Weak<T> {
-    /// Attempts to upgrade the `Weak` pointer to an [`Arc`], delaying
-    /// dropping of the inner value if successful.
-    ///
-    /// Returns [`None`] if the inner value has since been dropped.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use std::sync::Arc;
-    ///
-    /// let five = Arc::new(5);
-    ///
-    /// let weak_five = Arc::downgrade(&five);
-    ///
-    /// let strong_five: Option<Arc<_>> = weak_five.upgrade();
-    /// assert!(strong_five.is_some());
-    ///
-    /// // Destroy all strong pointers.
-    /// drop(strong_five);
-    /// drop(five);
-    ///
-    /// assert!(weak_five.upgrade().is_none());
-    /// ```
-    #[stable(feature = "arc_weak", since = "1.4.0")]
-    pub fn upgrade(&self) -> Option<Arc<T>> {
-        // We use a CAS loop to increment the strong count instead of a
-        // fetch_add as this function should never take the reference count
-        // from zero to one.
-        let inner = self.inner()?;
-
-        // Relaxed load because any write of 0 that we can observe
-        // leaves the field in a permanently zero state (so a
-        // "stale" read of 0 is fine), and any other value is
-        // confirmed via the CAS below.
-        let mut n = inner.strong.load(Relaxed);
-
-        loop {
-            if n == 0 {
-                return None;
-            }
-
-            // See comments in `Arc::clone` for why we do this (for `mem::forget`).
-            if n > MAX_REFCOUNT {
-                abort();
-            }
-
-            // Relaxed is fine for the failure case because we don't have any expectations about the new state.
-            // Acquire is necessary for the success case to synchronise with `Arc::new_cyclic`, when the inner
-            // value can be initialized after `Weak` references have already been created. In that case, we
-            // expect to observe the fully initialized value.
-            match inner.strong.compare_exchange_weak(n, n + 1, Acquire, Relaxed) {
-                Ok(_) => return Some(Arc::from_inner(self.ptr)), // null checked above
-                Err(old) => n = old,
-            }
-        }
-    }
-
-    /// Gets the number of strong (`Arc`) pointers pointing to this allocation.
-    ///
-    /// If `self` was created using [`Weak::new`], this will return 0.
-    #[stable(feature = "weak_counts", since = "1.41.0")]
-    pub fn strong_count(&self) -> usize {
-        if let Some(inner) = self.inner() { inner.strong.load(SeqCst) } else { 0 }
-    }
-
-    /// Gets an approximation of the number of `Weak` pointers pointing to this
-    /// allocation.
-    ///
-    /// If `self` was created using [`Weak::new`], or if there are no remaining
-    /// strong pointers, this will return 0.
-    ///
-    /// # Accuracy
-    ///
-    /// Due to implementation details, the returned value can be off by 1 in
-    /// either direction when other threads are manipulating any `Arc`s or
-    /// `Weak`s pointing to the same allocation.
-    #[stable(feature = "weak_counts", since = "1.41.0")]
-    pub fn weak_count(&self) -> usize {
-        self.inner()
-            .map(|inner| {
-                let weak = inner.weak.load(SeqCst);
-                let strong = inner.strong.load(SeqCst);
-                if strong == 0 {
-                    0
-                } else {
-                    // Since we observed that there was at least one strong pointer
-                    // after reading the weak count, we know that the implicit weak
-                    // reference (present whenever any strong references are alive)
-                    // was still around when we observed the weak count, and can
-                    // therefore safely subtract it.
-                    weak - 1
-                }
-            })
-            .unwrap_or(0)
-    }
-
-    /// Returns `None` when the pointer is dangling and there is no allocated `ArcInner`,
-    /// (i.e., when this `Weak` was created by `Weak::new`).
-    #[inline]
-    fn inner(&self) -> Option<WeakInner<'_>> {
-        if is_dangling(self.ptr.as_ptr()) {
-            None
-        } else {
-            // We are careful to *not* create a reference covering the "data" field, as
-            // the field may be mutated concurrently (for example, if the last `Arc`
-            // is dropped, the data field will be dropped in-place).
-            Some(unsafe {
-                let ptr = self.ptr.as_ptr();
-                WeakInner { strong: &(*ptr).strong, weak: &(*ptr).weak }
-            })
-        }
-    }
-
-    /// Returns `true` if the two `Weak`s point to the same allocation (similar to
-    /// [`ptr::eq`]), or if both don't point to any allocation
-    /// (because they were created with `Weak::new()`).
-    ///
-    /// # Notes
-    ///
-    /// Since this compares pointers it means that `Weak::new()` will equal each
-    /// other, even though they don't point to any allocation.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use std::sync::Arc;
-    ///
-    /// let first_rc = Arc::new(5);
-    /// let first = Arc::downgrade(&first_rc);
-    /// let second = Arc::downgrade(&first_rc);
-    ///
-    /// assert!(first.ptr_eq(&second));
-    ///
-    /// let third_rc = Arc::new(5);
-    /// let third = Arc::downgrade(&third_rc);
-    ///
-    /// assert!(!first.ptr_eq(&third));
-    /// ```
-    ///
-    /// Comparing `Weak::new`.
-    ///
-    /// ```
-    /// use std::sync::{Arc, Weak};
-    ///
-    /// let first = Weak::new();
-    /// let second = Weak::new();
-    /// assert!(first.ptr_eq(&second));
-    ///
-    /// let third_rc = Arc::new(());
-    /// let third = Arc::downgrade(&third_rc);
-    /// assert!(!first.ptr_eq(&third));
-    /// ```
-    ///
-    /// [`ptr::eq`]: core::ptr::eq
-    #[inline]
-    #[stable(feature = "weak_ptr_eq", since = "1.39.0")]
-    pub fn ptr_eq(&self, other: &Self) -> bool {
-        self.ptr.as_ptr() == other.ptr.as_ptr()
-    }
-}
-
-#[stable(feature = "arc_weak", since = "1.4.0")]
-impl<T: ?Sized> Clone for Weak<T> {
-    /// Makes a clone of the `Weak` pointer that points to the same allocation.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use std::sync::{Arc, Weak};
-    ///
-    /// let weak_five = Arc::downgrade(&Arc::new(5));
-    ///
-    /// let _ = Weak::clone(&weak_five);
-    /// ```
-    #[inline]
-    fn clone(&self) -> Weak<T> {
-        let inner = if let Some(inner) = self.inner() {
-            inner
-        } else {
-            return Weak { ptr: self.ptr };
-        };
-        // See comments in Arc::clone() for why this is relaxed.  This can use a
-        // fetch_add (ignoring the lock) because the weak count is only locked
-        // where are *no other* weak pointers in existence. (So we can't be
-        // running this code in that case).
-        let old_size = inner.weak.fetch_add(1, Relaxed);
-
-        // See comments in Arc::clone() for why we do this (for mem::forget).
-        if old_size > MAX_REFCOUNT {
-            abort();
-        }
-
-        Weak { ptr: self.ptr }
-    }
-}
-
-#[stable(feature = "downgraded_weak", since = "1.10.0")]
-impl<T> Default for Weak<T> {
-    /// Constructs a new `Weak<T>`, without allocating memory.
-    /// Calling [`upgrade`] on the return value always
-    /// gives [`None`].
-    ///
-    /// [`upgrade`]: Weak::upgrade
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use std::sync::Weak;
-    ///
-    /// let empty: Weak<i64> = Default::default();
-    /// assert!(empty.upgrade().is_none());
-    /// ```
-    fn default() -> Weak<T> {
-        Weak::new()
-    }
-}
-
-#[stable(feature = "arc_weak", since = "1.4.0")]
-impl<T: ?Sized> Drop for Weak<T> {
-    /// Drops the `Weak` pointer.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use std::sync::{Arc, Weak};
-    ///
-    /// struct Foo;
-    ///
-    /// impl Drop for Foo {
-    ///     fn drop(&mut self) {
-    ///         println!("dropped!");
-    ///     }
-    /// }
-    ///
-    /// let foo = Arc::new(Foo);
-    /// let weak_foo = Arc::downgrade(&foo);
-    /// let other_weak_foo = Weak::clone(&weak_foo);
-    ///
-    /// drop(weak_foo);   // Doesn't print anything
-    /// drop(foo);        // Prints "dropped!"
-    ///
-    /// assert!(other_weak_foo.upgrade().is_none());
-    /// ```
-    fn drop(&mut self) {
-        // If we find out that we were the last weak pointer, then its time to
-        // deallocate the data entirely. See the discussion in Arc::drop() about
-        // the memory orderings
-        //
-        // It's not necessary to check for the locked state here, because the
-        // weak count can only be locked if there was precisely one weak ref,
-        // meaning that drop could only subsequently run ON that remaining weak
-        // ref, which can only happen after the lock is released.
-        let inner = if let Some(inner) = self.inner() { inner } else { return };
-
-        if inner.weak.fetch_sub(1, Release) == 1 {
-            acquire!(inner.weak);
-            unsafe { Global.deallocate(self.ptr.cast(), Layout::for_value_raw(self.ptr.as_ptr())) }
         }
     }
 }
